% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_thesis.R
\name{georgetown_thesis}
\alias{georgetown_thesis}
\title{Creates an R Markdown PDF Thesis document}
\usage{
georgetown_thesis(toc = TRUE, toc_depth = 3)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying where table of contents should be created}

\item{toc_depth}{A positive integer}
}
\value{
A modified \code{pdf_document} based on the Georgetown Senior Thesis LaTeX
  template
}
\description{
This is a function called in output in the YAML of the driver Rmd file to specify using the Georgetown
Senior Thesis LaTeX template and cls files.
}
\examples{
\dontrun{
 output:
   georgetowntemplates::georgetown_thesis:
     toc: true
}
}

