% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctemplates.R
\docType{package}
\name{georgetowntemplates}
\alias{georgetowntemplates}
\alias{georgetowntemplates-package}
\title{georgetowntemplates: A package for creating documents using R Markdown}
\description{
The georgetowntemplates package provides three important functions:
georgetown_thesis, label, and ref.
}
\section{georgetown_thesis}{

Creates an R Markdown senior thesis template
}

\section{label}{

Adds a label to a figure/chemical reaction and includes the figure in the document
}

\section{ref}{

Allows references to be made to labelled objects inline in the document
}

